Problems 8-10 and adjustments
For problem 8 I adjusted the billing_schedule appropriately but I also
decided to slightly alter the make_invoices method in PolicyAccounting
to mark any previously stored invoices as deleted rather than deleting
them from the database. I believe that makes for a very simple solution
to someone wishing to change schedules in the middle of a policy. If the
invoices need to be deleted from the database, it would need to be done
manually (probably best keep record the old invoices anyway). Also
adjusted the return_account_balance method to only sum invoices not
marked as deleted.

For problem 9 I decided to create a new table, CanceledPolicy, in the
database that references a policy.  This is meant to hold all the
information one would need to know about a policy that has been
canceled. Creating this new table keeps the Policy table clean and free
of unnecessary data.

For problem 10 I made two web pages, one to receive the date and policy
number, and the other to display the data from the database. The latter
page shows payments as well as invoices.

BONUS:
Added a test that corresponds to problem 7, verifies that payments are
being processed correctly past the due date.

Consolidated the code in make invoices, the large if-elif block was
unnecessary. I also rearranged some of the code here to make sure we
weren't deleting invoices before checking that we had a valid
billing_schedule and that we weren't adding invoices to the database
despite having a bad billing_schedule.

Created an extra database relation between policies and payments, used
for displaying payments easily on policy view.